altoslib: Store saved state in version-independent format
Use AltosHashSet for AltosState so that AltosDroid doesn't lose
tracker information when the application is upgraded.

+++ altoslib/AltosConfigData.java
@@ -204,7 +204,7 @@ public class AltosConfigData implements Iterable<String> {
 
 		for (int i = 0; i < parts.length; i++) {
 			try {
-				r[i] = AltosLib.fromdec(parts[i]);
+				r[i] = (int) AltosLib.fromdec(parts[i]);
 			} catch (NumberFormatException n) {
 				r[i] = 0;
 			}
